pyproject.toml
[project]
name = "parkinsondiseaseclassifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "biopython>=1.84",
    "matplotlib>=3.9.2",
    "pandas>=2.2.3",
    "scikit-learn>=1.5.2",
    "seaborn>=0.13.2",
    "xgboost>=2.1.2",
]
